update vue examples code and simplify
