+++ examples/vue/vue-3/src/App.vue
@@ -1,5 +1,56 @@
-<script setup lang="ts">
-import DynamicallyRenderBuilderPage from '@/components/DynamicallyRenderBuilderPage.vue';
+<script lang="ts">
+import { RenderContent, getContent, isPreviewing } from '@builder.io/sdk-vue/vue3';
+import '@builder.io/sdk-vue/vue3/css';
+
+import HelloWorldComponent from './components/HelloWorld.vue';
+
+// Register your Builder components
+const REGISTERED_COMPONENTS = [
+  {
+    component: HelloWorldComponent,
+    name: 'Hello World',
+    canHaveChildren: true,
+    inputs: [
+      {
+        name: 'text',
+        type: 'string',
+        defaultValue: 'World',
+      },
+    ],
+  },
+];
+
+// TODO: enter your public API key
+const BUILDER_PUBLIC_API_KEY = 'f1a790f8c3204b3b8c5c1795aeac4660'; // ggignore
+
+export default {
+  name: 'DynamicallyRenderBuilderPage',
+  components: {
+    'builder-render-content': RenderContent,
+  },
+  data: () => ({
+    canShowContent: false,
+    content: null,
+    apiKey: BUILDER_PUBLIC_API_KEY,
+  }),
+  methods: {
+    getRegisteredComponents() {
+      return REGISTERED_COMPONENTS;
+    },
+  },
+  mounted() {
+    getContent({
+      model: 'page',
+      apiKey: BUILDER_PUBLIC_API_KEY,
+      userAttributes: {
+        urlPath: window.location.pathname,
+      },
+    }).then(res => {
+      this.content = res;
+      this.canShowContent = this.content || isPreviewing();
+    });
+  },
+};
 </script>
 
 <template>
@@ -7,7 +58,20 @@ import DynamicallyRenderBuilderPage from '@/components/DynamicallyRenderBuilderP
     <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
   </header>
   <div>
-    <DynamicallyRenderBuilderPage />
+    <div>Hello world from your Vue 3 project. Below is Builder Content:</div>
+    <div v-if="canShowContent">
+      <div>
+        page title:
+        {{ (content && content.data && content.data.title) || 'Unpublished' }}
+      </div>
+      <builder-render-content
+        model="page"
+        :content="content"
+        :api-key="apiKey"
+        :customComponents="getRegisteredComponents()"
+      />
+    </div>
+    <div v-else>Content not Found</div>
   </div>
 </template>
 
@@ -17,36 +81,5 @@ import DynamicallyRenderBuilderPage from '@/components/DynamicallyRenderBuilderP
   padding: 2rem;
 
   font-weight: normal;
-}
-
-header {
-  line-height: 1.5;
-  max-height: 100vh;
-}
-
-.logo {
-  display: block;
-  margin: 0 auto 2rem;
-}
-
-@media (min-width: 1024px) {
-  body {
-    display: flex;
-    place-items: center;
-  }
-
-  #app {
-    padding: 0 2rem;
-  }
-
-  header {
-    display: flex;
-    place-items: center;
-    justify-content: center;
-  }
-
-  .logo {
-    margin: 0 2rem 0 0;
-  }
 }
 </style>
